backend/requirements.txt
# Flask and web framework (keeping for compatibility)
flask==2.3.3
flask-cors==4.0.0
werkzeug==2.3.7
gunicorn==21.2.0
trustcall

# FastAPI - Performance optimized
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
starlette>=0.38.6
pydantic>=2.7.4
pydantic[email]>=2.7.4

# Database with connection pooling
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Authentication - Optimized bcrypt
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# API clients
notion-client==2.0.0
google-api-python-client==2.107.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.0

# Date handling
python-dateutil==2.8.2
pytz==2023.3
dateparser==1.1.8

# LangChain and LLM components
langchain>=0.1.0
langchain-community>=0.0.29
langchain-core>=0.2.38
langchain-anthropic>=0.2.0
langchain-openai>=0.1.0
langgraph>=0.3.27
langgraph-supervisor


# AI/ML
anthropic>=0.23.0,<1.0
openai>=1.10.0,<2.0.0

# Rate limiting and retry logic
ratelimit==2.2.1
backoff==2.2.1
tenacity==8.2.3

# HTML parsing
beautifulsoup4==4.12.2

# Environment variables
python-dotenv==1.0.0

# Utilities
tqdm==4.66.1
numpy==1.26.2
httpx>=0.25.2

# Performance and optimization - FIXED VERSION CONFLICT
python-multipart==0.0.6
orjson>=3.10.1  # ✅ Updated to satisfy langgraph-sdk requirement

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
bandit>=1.7.5
safety>=2.3.0

# Supabase integration (using custom HTTP client to avoid conflicts)
